+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.6.0
+
+* tbd
+
+## 0.5.6
+
+* Cache test run performance and run slowest tests first to maximiz worker
+  utilization
+* Update to jsdom 6.5.0
+
+## 0.5.5
+
+* Improve failure stack traces.
+* Fix syntax error reporting.
+* Add `--watch` option (#472).
+
 ## 0.5.2
 
 * Fixed a bug with syntax errors in test files (#487).